tidying

+++ advanced/sr_gui_hand_calibration/src/sr_gui_hand_calibration/sr_hand_calibration_model.py
@@ -195,20 +195,22 @@ class JointCalibration(QTreeWidgetItem):
         self.package_path = package_path
         self.multiplot_processes = []
         REMOTE_PLOTJUGGLER_VARIABLES = ['SERVER_IP', 'SERVER_USERNAME', 'CONTAINER_NAME']
-        self.local_plotjuggler = True
+        self._nuc_ssh_key_path = "/home/user/.ssh/reverse_ssh_id_rsa"
+        self._local_plotjuggler = True
         self._server_ip = None
         self._server_username = None
         self._container_name = None
-        rospy.loginfo("###############################")
         if any(var in os.environ for var in REMOTE_PLOTJUGGLER_VARIABLES):
             if all(var in os.environ for var in REMOTE_PLOTJUGGLER_VARIABLES):
-                self.local_plotjuggler = False
+                self._local_plotjuggler = False
                 self._server_ip = os.environ.get('SERVER_IP')
                 self._server_username = os.environ.get('SERVER_USERNAME')
                 self._container_name = os.environ.get('CONTAINER_NAME')
             else:
-                rospy.logwarn("Some but not all remote plotjuggler variables are set. This means there has been a "\
-                              "deployment error. Please contact shadow")
+                rospy.logerr("Some but not all remote plotjuggler variables are set. "\
+                             "This means there has been a deployment error. Please contact "\
+                             "software@shadowrobot.com")
+                rospy.logwarn("Defaulting to using plotjuggler locally")
 
         if not isinstance(self.joint_name, list):
             QTreeWidgetItem.__init__(
@@ -238,82 +240,76 @@ class JointCalibration(QTreeWidgetItem):
         tree_widget.addTopLevelItem(self)
         self.timer.timeout.connect(self.update_joint_pos)
 
-    def _create_ssh_client(self, server, port, user):
-        k = paramiko.RSAKey.from_private_key_file("/home/user/.ssh/reverse_ssh_id_rsa")
+    def _create_ssh_client(self):
+        try:
+            key = paramiko.RSAKey.from_private_key_file(self._nuc_ssh_key_path)
+        except FileNotFoundError as exception:
+            rospy.logerr(f"Failed to load SSH key - {exception}")
+            return
         client = paramiko.SSHClient()
         client.load_system_host_keys()
         client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-        client.connect(server, port=port, username=user, pkey=k)
+        try:
+            client.connect(self._server_ip, port=22, username=self._server_username, pkey=key)
+        except (NoValidConnectionsError, BadHostKeyException, AuthenticationException,
+                SSHException, socket.error) as exception:
+            rospy.logerr(f"Failed to SSH back to server - {exception}")
         return client
 
-    def ssh_command(self, ip, username, container_name, command):
-        k = paramiko.RSAKey.from_private_key_file("/home/user/.ssh/reverse_ssh_id_rsa")
-        client = paramiko.SSHClient()
-        client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-        try:
-            client.connect(self._server_ip, 22, self._server_username, pkey=k)
-            _, stdout, stderr = client.exec_command(command)
-            rospy.logwarn(f"cmdwas: {command}")
-            std_out = stdout.read()
-            if std_out:
-                rospy.logwarn(f"stdout: {std_out}")
-            std_err = stderr.read()
-            if std_err:
-                rospy.logwarn(f"stderr: {std_err}")
-            arm_serial_number = stdout.readline()
-            client.close()
-        except NoValidConnectionsError as exception:
-            ssh_exception_message = f"Failed to SSH into arm - {exception}"
-        except (BadHostKeyException, AuthenticationException, SSHException, socket.error) as exception:
-            ssh_exception_message = f"Failed to SSH into arm - {exception}"
-
-    def get_server_info(self):
-        with open('/tmp/server_username', 'r') as f:
-            server_username = f.readline().strip('\n')
-            server_containername = f.readline().strip('\n')
-        return server_username, server_containername
+    def _ssh_command(self, command):
+        client = self._create_ssh_client()
+        _, stdout, stderr = client.exec_command(command)
+        std_out = stdout.read()
+        if std_out:
+            rospy.loginfo(f"SSH command response: {std_out}")
+        std_err = stderr.read()
+        if std_err:
+            rospy.logwarn(f"SSH stderr: {std_err}")
+            rospy.logwarn(f"SSH command was: {command}")
+        client.close()
 
     def _wrap_in_docker_exec(self, command, detach=False):
         if detach:
             return f"docker exec -d {self._container_name} bash -c '{command}'"
         return f"docker exec {self._container_name} bash -c '{command}'"
 
-    def _start_remote_plotjuggler(self, command):
-        command = [x.replace("'", "\'") for x in command]
-        run_script_command = f"source /home/user/projects/shadow_robot/base/devel/setup.bash"
-        tmp_nuc_script_path = "/tmp/ssh_start_plotjuggler.sh"
-        command_str = " ".join(command)
-        command_str = f"{run_script_command} && {command_str}"
-        rospy.logerr(f"writing {command_str} to file {tmp_nuc_script_path}")
+    @staticmethod
+    def _prepend_source_ros(command):
+        return f"source /home/user/projects/shadow_robot/base/devel/setup.bash && {command}"
+
+    def _start_remote_plotjuggler(self, rosrun_command):
+        # Escape quotes
+        # rosrun_command = [x.replace("'", "\'") for x in rosrun_command]
+        tmp_script_path = "/tmp/ssh_start_plotjuggler.sh"
+        rosrun_command_str = self._prepend_source_ros(" ".join(rosrun_command))
         try:
-            with open(tmp_nuc_script_path, "w+", encoding="ASCII") as tmp_file:
-                tmp_file.write(command_str)
+            with open(tmp_script_path, "w+", encoding="ASCII") as tmp_file:
+                tmp_file.write(rosrun_command_str)
         except Exception:
-            rospy.logerr("Failed to nucssssss ation file: {}".format(tmp_nuc_script_path))
+            rospy.logerr(f"Failed to write {rosrun_command_str} to {tmp_script_path}")
             return
-        self._send_file(tmp_nuc_script_path, tmp_nuc_script_path)
-        copy_script_command = f"docker cp /tmp/ssh_start_plotjuggler.sh {self._container_name}:/tmp/ssh_start_plotjuggler.sh"
-        enable_script_command = self._wrap_in_docker_exec("sudo chmod +x /tmp/ssh_start_plotjuggler.sh")
+        self._send_file(local_path=tmp_script_path, remote_path=tmp_script_path)
+        copy_script_command = f"docker cp {tmp_script_path} {self._container_name}:{tmp_script_path}"
+        enable_script_command = self._wrap_in_docker_exec(f"sudo chmod +x {tmp_script_path}")
         run_script_command = self._wrap_in_docker_exec("cd /tmp && ./ssh_start_plotjuggler.sh", detach=True)
         for command in [copy_script_command, enable_script_command, run_script_command]:
-            self.ssh_command(self._server_ip, self._server_username, self._container_name, command)
+            self._ssh_command(command)
 
     def _send_file(self, local_path, remote_path):
-        rospy.logwarn(f"Sending file {local_path} to {remote_path}")
-        ssh_client = self._create_ssh_client(self._server_ip, 22, self._server_username)
+        rospy.loginfo(f"Sending template file {local_path} to {remote_path}")
+        ssh_client = self._create_ssh_client()
         scp = SCPClient(ssh_client.get_transport())
         scp.put(local_path, remote_path)
         ssh_client.close()
-        rospy.logwarn(f"Sent file {local_path} to {remote_path}")
 
     def _send_template(self, template):
-        self._send_file(template, '/tmp/tmp_plot.xml')
-        copy_command = f"docker cp /tmp/tmp_plot.xml {self._container_name}:/tmp/tmp_plot.xml"
-        self.ssh_command(self._server_ip, self._server_username, self._container_name, copy_command)
-        source_command = f"source /home/user/projects/shadow_robot/base/devel/setup.bash"
-        move_command = "mv /tmp/tmp_plot.xml $(rospack find sr_gui_hand_calibration)/resource/tmp_plot.xml"
-        full_command = f"docker exec {self._container_name} bash -c '{source_command} && {move_command}'"
-        self.ssh_command(self._server_ip, self._server_username, self._container_name, full_command)
+        tmp_template_path = "/tmp/tmp_plot.xml"
+        self._send_file(local_path=template, remote_path=tmp_template_path)
+        docker_cp_command = f"docker cp /tmp/tmp_plot.xml {self._container_name}:{tmp_template_path}"
+        self._ssh_command(docker_cp_command)
+        move_command = f"mv {tmp_template_path} $(rospack find sr_gui_hand_calibration)/resource/tmp_plot.xml"
+        move_command = self._prepend_source_ros(move_command)
+        self._ssh_command(self._wrap_in_docker_exec(move_command))
 
     def plot_raw_button_clicked(self):
         temporary_file_name = "{}/resource/tmp_plot.xml".format(self.package_path)
@@ -352,8 +348,6 @@ class JointCalibration(QTreeWidgetItem):
         if len(hand_serial_side_dict) > 1:
             rospy.logerr("More than one hand detected, please only connect one hand.")
             return
-        # hand_finder = HandFinder()
-        # prefix = hand_finder.get_available_prefix()
         prefix = f"{list(hand_serial_side_dict.values())[0]}_"
         if prefix == 'lh_':
             replace_list.append(['/rh/', '/lh/'])
@@ -365,7 +359,7 @@ class JointCalibration(QTreeWidgetItem):
         except Exception:
             rospy.logerr("Failed to write temportary multiplot configuration file: {}".format(temporary_file_name))
             return
-        if not self.local_plotjuggler:
+        if not self._local_plotjuggler:
             self._send_template(temporary_file_name)
             self._start_remote_plotjuggler(process)
         else:
@@ -445,8 +439,9 @@ class JointCalibration(QTreeWidgetItem):
 
     def on_close(self):
         self.timer.stop()
-        for process in self.multiplot_processes:
-            process.terminate()
+        if not self._local_plotjuggler:
+            for process in self.multiplot_processes:
+                process.terminate()
 
 
 class FingerCalibration(QTreeWidgetItem):
